Move process logger getter and setter to the Communicator
String and Int versions are now also static  properties of the communicator
